Added Test cases. Fixed null issue
Added test fighters and made multiple test cases to see that everything is working.
Also, changed out all instancesof "null" to insteaduse a harmless nullFighter. We check this and if they are a no show, we can safely deploy a bye. If we have two no shows against each other, we can skip that match entirely.